requirements.txt
# requirements.txt for Python 3.8
# AI-PumpDiag - Dynamogram Classification System

# GUI Framework
PyQt5==5.15.7

# Deep Learning & AI
torch==1.13.1
torchvision==0.14.1

# Image Processing
Pillow==9.3.0

# PDF Processing
PyMuPDF==1.21.1

# Scientific Computing
numpy==1.21.6
scipy==1.9.3

# Machine Learning & Metrics
scikit-learn==1.1.3

# Data Visualization
matplotlib==3.5.3
seaborn==0.11.2

# Progress Bars (optional, for CLI usage)
tqdm==4.64.1

# JSON handling (built-in, but ensuring compatibility)
# json - built-in module

# Path handling (built-in in Python 3.8+)
# pathlib - built-in module

# System utilities
psutil==5.9.4

# Hash functions (built-in)
# hashlib - built-in module

# Threading (built-in)
# threading - built-in module

# Date/time handling (built-in)
# datetime - built-in module

# File operations (built-in)
# os, sys, io - built-in modules

# Optional: CUDA support (uncomment if you have NVIDIA GPU)
# torch==1.13.1+cu117 -f https://download.pytorch.org/whl/cu117/torch_stable.html
# torchvision==0.14.1+cu117 -f https://download.pytorch.org/whl/cu117/torch_stable.html

# Optional: CPU-only PyTorch (if you want to force CPU usage)
# torch==1.13.1+cpu -f https://download.pytorch.org/whl/cpu/torch_stable.html
# torchvision==0.14.1+cpu -f https://download.pytorch.org/whl/cpu/torch_stable.html

# Development dependencies (optional)
# pytest==7.2.0
# black==22.10.0
# flake8==5.0.4

# Note: Versions are specifically chosen for Python 3.8 compatibility
# These versions are tested and stable for the dynamogram classification system